11.1.21 better terminology for translucency
color translucent [128-256] or [0.0 - 1.0]
actually implemented: three cutoffs; 0.5 (128), 0.75 (192), and 0.825 (224)

+++ Jmol/src/org/jmol/viewer/Dipole.java
@@ -90,7 +90,7 @@ class Dipole {
   void setProperty(String propertyName, Object value, BitSet bsSelected) {
   }
 
-  void setTranslucent(boolean isTranslucent, int iLevel) {
+  void setTranslucent(boolean isTranslucent, float iLevel) {
     colix = Graphics3D.getColixTranslucent(colix, isTranslucent, iLevel);
   }
